Changes made in User, Item, Message and Topic classes in order to fix compareTo method.
Now all tests (except 20) are running properly on windows, and there is no problem with the system time anymore. Lend-me is able to notice the difference between the creation time of the objects (even if they were created one just after the other and this difference is almost negligible), using nanoseconds precision.