Task 7: Make an Observer to notify the interested users that their desired item is available.
Well, I did it. But not in the usual way. As the action when the item is available is done by the observed, not by the observer, I had to make an interface that mirrored this behavior.

Using generics I created an interface like this:

public interface InterestedOn<K> {

	public void warn...(K target, InterestedOn<K> interested) {
		...
	}
	
}

This way it forces the implementor of this class to warn another guy that implements the interface, and thus interest can be mutual.