Added authentication Facebook client support.
It already works and retrieves user address just fine.
The only thing that is still needed is to make the client side redirect the browser to the link that is
received by the login() method, so the user can login and then go back to our website and thus our system must call the authenticate(access_token) method that actually checks whether the user really logged in facebook. For testing purposes, I've made the return of the authenticate method to be the address of the user.